news outlets/requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
cymem==2.0.7
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fake-headers==1.0.2
Flask==2.2.2
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.9
numpy==1.24.1
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pathy==0.10.1
pip==22.3.1
plotly==5.11.0
preshed==3.0.8
pycparser==2.21
pydantic==1.10.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
selenium==4.7.2
setuptools==65.5.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
tenacity==8.1.0
thinc==8.1.6
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.4.0
undetected-chromedriver==3.1.7
urllib3==1.26.13
wasabi==0.10.1
webdriver-manager==3.8.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
wsproto==1.2.0
zipp==3.11.0
